Cleanup tcp
* Got rid of unwanted class hierarcy for IOSelector; now there is one selector that does reading/writing

* Renamed IOSelector to IOReactor since IOSelector is not correct name. It wraps a Selector, but it is not
a selector

The reason for this cleanup is that:

* the code is simpler to understand (less hierarchies without value)

* the code becomes easier to tweak for performance. A few of the experiments planned:

  1) have a single thread that spins on socketchannel and takes care of all io (no selectors)

  2) combine input with output 'selectors'. So instead of 3 threads doing input and 3 threads doing output;
     have 3 threads that do input + output.
