+++ hazelcast/src/test/java/com/hazelcast/nio/tcp/iobalancer/LoadTrackerTest.java
@@ -17,8 +17,7 @@
 package com.hazelcast.nio.tcp.iobalancer;
 
 import com.hazelcast.logging.ILogger;
-import com.hazelcast.logging.LoggingService;
-import com.hazelcast.nio.tcp.AbstractIOSelector;
+import com.hazelcast.nio.tcp.IOReactorImp;
 import com.hazelcast.nio.tcp.MigratableHandler;
 import com.hazelcast.test.HazelcastParallelClassRunner;
 import com.hazelcast.test.annotation.QuickTest;
@@ -35,17 +34,17 @@ import static org.mockito.Mockito.when;
 @Category(QuickTest.class)
 public class LoadTrackerTest {
 
-    private AbstractIOSelector selector1;
-    private AbstractIOSelector selector2;
+    private IOReactorImp selector1;
+    private IOReactorImp selector2;
 
-    private AbstractIOSelector[] selectors;
+    private IOReactorImp[] selectors;
     private LoadTracker loadTracker;
 
     @Before
     public void setUp() {
-        selector1 = mock(AbstractIOSelector.class);
-        selector2 = mock(AbstractIOSelector.class);
-        selectors = new AbstractIOSelector[]{selector1, selector2};
+        selector1 = mock(IOReactorImp.class);
+        selector2 = mock(IOReactorImp.class);
+        selectors = new IOReactorImp[]{selector1, selector2};
 
         ILogger logger = mock(ILogger.class);
         when(logger.isFinestEnabled()).thenReturn(true);
